# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quartermaster API",
    "description": "Self-hosted kitchen inventory management.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/reminders": {
      "get": {
        "tags": [
          "reminders"
        ],
        "operationId": "reminders_list",
        "parameters": [
          {
            "name": "after_fire_at",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "after_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderListResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reminders/{id}/ack": {
      "post": {
        "tags": [
          "reminders"
        ],
        "operationId": "reminders_ack",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reminders/{id}/present": {
      "post": {
        "tags": [
          "reminders"
        ],
        "operationId": "reminders_present",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable error code, machine-readable."
          },
          "message": {
            "type": "string",
            "description": "Human-readable description. Not localised; clients should prefer `code`."
          },
          "unrestorable_ids": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Present on `code = batch_not_restorable` when the failure came from a\nbulk call — lists the specific batches the caller needs to look at.\nAbsent on single-batch restore (the path already names the batch)."
          }
        }
      },
...
      "ReminderDto": {
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "body",
          "fire_at",
          "household_timezone",
          "household_fire_local_at",
          "batch_id",
          "product_id",
          "location_id"
        ],
        "properties": {
          "acked_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "batch_id": {
            "type": "string",
            "format": "uuid"
          },
          "body": {
            "type": "string"
          },
          "expires_on": {
            "type": [
              "string",
              "null"
            ]
          },
          "fire_at": {
            "type": "string"
          },
          "household_fire_local_at": {
            "type": "string"
          },
          "household_timezone": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/ReminderKind"
          },
          "location_id": {
            "type": "string",
            "format": "uuid"
          },
          "presented_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "product_id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "ReminderKind": {
        "type": "string",
        "enum": [
          "expiry"
        ]
      },
      "ReminderListResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReminderDto"
            }
          },
          "next_after_fire_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "next_after_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      }
    }
  },
...
}